Day 20 final
input algo bit 0 is a 1. That means on every odd step, *every* bit in the infinite grid will be "on".

The only way this ever works is if the 512th bit is always a 0 when the 1st bit is 1, so we're not gonna even check that.

This means that if we're on an odd step, and we're looking at bits outside of our input image, we need to pretend those bits are on, not off.