Day 24
Manually trying to run the asm code.

It's a push/pop series. Every time x might have a value that could be equal to w, we assume it is, as that's the only way z is reduced.

It's incorrect, but most likely because my math is wrong. I'll run it again.