Found an algorithm to automate the manual steps
Very simple, but zero generalization. Basically runs through the steps 18 commands at a time, ignores all the actual commands and variables.

Check operand of command 4.
If it's 1, saves the operand of command 15 (and current idx)
if it's not (it's 26), pop the last operand, and add it to operand 5.
The formula for the popped index and current index is

cur_idx=popped_idx+v